EN6-82 Refactor content-attributes components to avoid dependency cycle
