+++ src/ui/edit-content/content-attributes/ContentAttributes.js
@@ -2,46 +2,11 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import {
   FieldArray,
-  fieldArrayFieldsPropTypes,
 } from 'redux-form';
 import { Tabs, Tab } from 'patternfly-react';
-
-import AttributeField from 'ui/edit-content/content-attributes/AttributeField';
-import { TYPE_TIMESTAMP } from 'state/content-type/const';
 import { FormattedMessage } from 'react-intl';
 
-const AttributeFields = ({ attributes, fields, settings }) => {
-  if (fields.length < attributes.length) {
-    // initialize fields with values from attributes prop through `.push()` method
-    // as it cannot be set directly from props
-    attributes.forEach((attr) => {
-      const {
-        value, elements, compositeelements, type,
-      } = attr;
-      fields.push(
-        (type === TYPE_TIMESTAMP && {})
-        || (elements && { elements })
-        || (compositeelements && { compositeelements })
-        || value,
-      );
-    });
-  }
-
-  return fields.map((attrRef, idx) => (
-    <AttributeField
-      key={attributes[idx].code}
-      name={attrRef}
-      attribute={attributes[idx]}
-      settings={settings}
-    />
-  ));
-};
-
-AttributeFields.propTypes = {
-  attributes: PropTypes.arrayOf(PropTypes.object).isRequired,
-  fields: PropTypes.shape(fieldArrayFieldsPropTypes).isRequired,
-  settings: PropTypes.shape({}).isRequired,
-};
+import AttributeFields from 'ui/edit-content/content-attributes/AttributeFields';
 
 class ContentAttributes extends Component {
   componentDidMount() {
@@ -53,7 +18,7 @@ class ContentAttributes extends Component {
     const { attributes, settings } = this.props;
 
     return (
-      <Tabs defaultActiveKey={1} animation={false}>
+      <Tabs defaultActiveKey={1} animation={false} id="content-attributes-tabs">
         <Tab eventKey={1} title="English">
           <FieldArray
             data-test-id="edit-content-content-attributes-field-array"
