+++ src/ui/edit-content/content-attributes/__tests__/AttributeField.test.js
@@ -8,9 +8,6 @@ import {
   TYPE_DATE,
   TYPE_ENUMERATOR,
   TYPE_ENUMERATOR_MAP,
-  TYPE_COMPOSITE,
-  TYPE_LIST,
-  TYPE_MONOLIST,
   TYPE_NUMBER,
   TYPE_THREESTATE,
   TYPE_TIMESTAMP,
@@ -32,9 +29,6 @@ const attributeFields = {
   [TYPE_DATE]: 'DateAttributeField',
   [TYPE_ENUMERATOR]: 'EnumeratorAttributeField',
   [TYPE_ENUMERATOR_MAP]: 'EnumeratorMapAttributeField',
-  [TYPE_COMPOSITE]: 'CompositeAttributeField',
-  [TYPE_LIST]: 'ListAttributeField',
-  [TYPE_MONOLIST]: 'MonolistAttributeField',
   [TYPE_NUMBER]: 'NumberAttributeField',
   [TYPE_THREESTATE]: 'ThreeStateAttributeField',
   [TYPE_TIMESTAMP]: 'TimestampAttributeField',
